requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
pillow==10.1.0
python-dotenv==1.0.0